User controller에서 인증 관련 부분들을 auth Contoroller로 변경
